refactor(core): co-locate field definition methods
