+++ packages/sanity/src/core/studio/components/navbar/search/definitions/operators/operatorTypes.ts
@@ -1,5 +1,5 @@
 import {ComponentType} from 'react'
-import {SearchFieldDefinition} from '../../types'
+import type {SearchFieldDefinition} from '../fields'
 
 /**
  * @alpha
