--- packages/sanity/src/core/studio/components/navbar/search/utils/createFieldDefinitionDictionary.ts
@@ -1,10 +1,0 @@
-import type {SearchFieldDefinition, SearchFieldDefinitionDictionary} from '../types'
-
-export function createFieldDefinitionDictionary(
-  fieldDefinitions: SearchFieldDefinition[]
-): SearchFieldDefinitionDictionary {
-  return fieldDefinitions.reduce<SearchFieldDefinitionDictionary>((acc, val) => {
-    acc[val.id] = val
-    return acc
-  }, {})
-}
