--- packages/sanity/src/core/studio/components/navbar/search/utils/createFilterDefinitionDictionary.ts
@@ -1,10 +1,0 @@
-import type {SearchFilterDefinition, SearchFilterDefinitionDictionary} from '../definitions/filters'
-
-export function createFilterDefinitionDictionary(
-  filterDefinitions: SearchFilterDefinition[]
-): SearchFilterDefinitionDictionary {
-  return filterDefinitions.reduce<SearchFilterDefinitionDictionary>((acc, val) => {
-    acc[val.name] = val
-    return acc
-  }, {})
-}
