--- packages/sanity/src/core/studio/components/navbar/search/utils/createOperatorDefinitionDictionary.ts
@@ -1,13 +1,0 @@
-import type {
-  SearchOperatorDefinition,
-  SearchOperatorDefinitionDictionary,
-} from '../definitions/operators'
-
-export function createOperatorDefinitionDictionary(
-  operatorDefinitions: SearchOperatorDefinition[]
-): SearchOperatorDefinitionDictionary {
-  return operatorDefinitions.reduce<SearchOperatorDefinitionDictionary>((acc, val) => {
-    acc[val.type] = val
-    return acc
-  }, {})
-}
