+++ packages/sanity/src/core/studio/components/navbar/search/utils/filterUtils.ts
@@ -2,6 +2,7 @@ import intersection from 'lodash/intersection'
 import isEmpty from 'lodash/isEmpty'
 import type {SearchableType} from '../../../../../search'
 import {isNonNullable} from '../../../../../util'
+import type {SearchFieldDefinition, SearchFieldDefinitionDictionary} from '../definitions/fields'
 import {
   getFilterDefinition,
   type OperatorItem,
@@ -9,7 +10,7 @@ import {
   type SearchFilterDefinitionDictionary,
 } from '../definitions/filters'
 import {getOperatorDefinition, SearchOperatorDefinitionDictionary} from '../definitions/operators'
-import type {SearchFieldDefinition, SearchFieldDefinitionDictionary, SearchFilter} from '../types'
+import type {SearchFilter} from '../types'
 
 export function buildSearchFilter(
   filterDefinition: SearchFilterDefinition,
